+++ src/components/MapComponent.vue
@@ -382,6 +382,8 @@ export default {
     &.map-component--minimized
       .handle
         display none!important
+    .handle
+      position absolute
     .handle-mr, .handle-ml
       top 0
       height 100%
@@ -392,13 +394,15 @@ export default {
       display block!important
       z-index 1001
     .handle-mr
+      cursor e-resize
       right 0
       border-left 1px solid #616161
     .handle-ml
+      cursor w-resize
       left 0
       border-right 1px solid #616161
     .handle-tm, .handle-bm
-      left 5px
+      left 0
       width 100%
       margin-top 0
       height 4px
@@ -407,9 +411,11 @@ export default {
       display block!important
       z-index 1001
     .handle-tm
+      cursor n-resize
       top 0
       border-bottom 1px solid #616161
     .handle-bm
+      cursor s-resize
       bottom 0
       border-top 1px solid #616161
     .handle-tl, .handle-bl, .handle-br, .handle-tr
@@ -421,15 +427,19 @@ export default {
       display block!important
       z-index 1002
     .handle-tl
+      cursor nw-resize
       left 0
       top 0
     .handle-bl
+      cursor sw-resize
       left 0
       bottom 0
     .handle-br
+      cursor nwse-resize
       right 0
       bottom 0
     .handle-tr
+      cursor nesw-resize
       right 0
       top 0
 </style>